pyproject.toml
[tool.poetry]
name = "zanjabil-backend"
version = "0.1.0"
description = ""
authors = ["Ruslan Korneev <ruslan@kortechs.io>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
